# backend/application/src/main/resources/api/v1/specs/openapi.yaml
openapi: 3.0.3
info:
  title: Report Constructor API
  version: 1.0.0
  description: |
    Report Constructor — turns a marketing-campaign Google Sheet into a generated
    Google Slides deck (campaign/tactic placeholder resolution, line-item matching,
    Claude-generated copy, native Sheets-linked charts). Migrated from the legacy
    PHP app onto the AI Digital locked stack.

    Contract-first: backend interfaces and frontend types are generated from this
    file. Schemas are versioned explicitly in their names (UserV1, SheetReadResultV1).
    Enums are standalone `<Noun>V1` `$ref` schemas. Every path key carries the
    literal `/api/v1/` prefix (the generator ignores `servers[].url`).

    Auth: Clerk SSO only — every operation requires a Bearer JWT validated against
    the Clerk JWKS. The backend reads identity from the `user_id` claim and email
    from the `email` claim. Sheet/line-item/placeholder endpoints are stateless per
    request (the client holds the 2-D sheet rows in wizard state and re-POSTs them);
    only report generation is persisted (an async job polled by id).
servers:
  - url: /api/v1
    description: V1 base path (DOCUMENTATION ONLY — the openapi-generator-maven-plugin
      does NOT prepend this to generated Spring routes. Path keys below carry the
      `/api/v1/` prefix verbatim. See canonical-openapi-rules.md → "/api/v1 prefix").
...
paths:
...
  /api/v1/sheets/read:
    post:
      tags: [ DataSource ]
      operationId: readSheet
      summary: Read one tab of a Google Sheet (A1:Z) as the service account
      description: |
        Reads a single tab (range `'<tab>'!A1:Z`, 26-column cap) and returns its
        raw 2-D rows plus a small preview and tab metadata. The service account must
        have read access to the spreadsheet (shared with it or link-readable). When
        the requested tab does not exist, returns 200 with `ok=false` and
        `error="tab_not_found"` so the client can apply its Estimates→Proposal fallback.
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SheetReadRequestV1"
      responses:
        "200":
          description: Tab read result (check `ok`/`error` for per-tab outcome).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetReadResultV1"
        "400":
          $ref: "#/components/responses/BadRequestV1"
        "401":
          $ref: "#/components/responses/UnauthorizedV1"
        "403":
          $ref: "#/components/responses/ForbiddenV1"
        "500":
          $ref: "#/components/responses/InternalServerErrorV1"
  /api/v1/google/status:
    get:
      tags: [ DataSource ]
      operationId: getGoogleConnectionStatus
      summary: Google connectivity for the signed-in user (service account + mock mode)
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: Connection status for sheet read and deck generation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoogleConnectionStatusV1"
        "401":
          $ref: "#/components/responses/UnauthorizedV1"
        "403":
          $ref: "#/components/responses/ForbiddenV1"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedV1:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorV1"
          examples:
            missingToken:
              value: { code: "C005", message: "Authentication required", timestamp: "2026-05-24T10:00:00Z", correlationId: "abc-123" }
    ForbiddenV1:
      description: Authenticated but lacking permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorV1"
    BadRequestV1:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorV1"
...
    InternalServerErrorV1:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorV1"
  schemas:
...
    ApiErrorV1:
      type: object
      required: [ code, message, timestamp ]
      properties:
        code: { type: string, description: "Machine-readable error code" }
        message: { type: string, description: "Human-readable message" }
        timestamp: { type: string, format: date-time }
        correlationId: { type: string }
        parameters:
          type: array
          items: { $ref: "#/components/schemas/ValidationParameterV1" }
    ValidationParameterV1:
      type: object
      required: [ code, value ]
      properties:
        code: { type: string }
        value: { type: string }

    # ── Enums (standalone $ref schemas — never inline) ──────────────────────────
...
    GoogleConnectionStatusV1:
      type: object
      required: [ connected, email, scopes, connectUrl, mockMode ]
      properties:
        connected: { type: boolean }
        email: { type: string }
        scopes:
          type: array
          items: { type: string }
        connectUrl: { type: string }
        mockMode: { type: boolean }
    SheetReadRequestV1:
      type: object
      required: [ url, tab ]
      properties:
        url: { type: string, description: "Full Google Sheets URL (the spreadsheet id is parsed out)." }
        tab: { type: string, description: "Exact tab name, e.g. Proposal, Audience&Inventory, Estimates, Geo, Basic." }
    SheetReadResultV1:
      type: object
      required: [ ok, tab, rows, cols, tabs, headers, preview, rawRows ]
      properties:
        ok: { type: boolean, description: "True when the tab was read; false with `error` set otherwise." }
        error: { type: string, description: "Error code when ok=false, e.g. tab_not_found." }
        sheetId: { type: string }
        title: { type: string, description: "Spreadsheet title." }
        tab: { type: string, description: "Tab that was read." }
        tabs:
          type: array
          description: All tab titles in the spreadsheet.
          items: { type: string }
        rows: { type: integer, format: int32 }
        cols: { type: integer, format: int32 }
        headers:
          type: array
          description: First row (header cells).
          items: { type: string }
        preview:
          type: array
          description: First few data rows for display.
          items:
            type: array
            items: { type: string }
        rawRows:
          type: array
          description: All rows of the tab (ragged-tolerant 2-D grid, A1:Z).
          items:
            type: array
            items: { type: string }

    # ── LineItemMatching ─────────────────────────────────────────────────────────
...